requirements.txt
fastapi
pyasic==0.62.2
uvicorn
pyyaml
redis 
fastapi-utils
setuptools